requirements.txt
Django==2.2.10
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.9.4
drf-yasg==1.17.1
psycopg2-binary==2.8.3
Faker==8.10.1
factory-boy==3.2.0
pytest-mock==3.6.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
